Fix deadlocking UI test case
The test did not wait long enough for a popup to appear. After catching
the timeout exception, it tried to delete the project as part of the
test cleanup. Deleting the project was blocked by the import wizard
that was still running, waiting for user input that would never come.

Increased the timout so the test passes on slower machines. Stop all
jobs before trying to cleanup the project.

+review BUILDSHIP-7

Signed-off-by: Stefan Oehme <st.oehme@gmail.com>