Make synchronization logic consistent
All synchronization steps are run regardless whether the project existed
before or not. This makes Buildships behavior more predictable to the
user.

Also, all the Java related steps are run eagerly instead of waiting for
them to be run by the ClasspathContainerInitializer. This has several
benefits

- the logic is easier to follow
- the tests only have to wait for one job, making them faster
- the ClasspathContainerInitializer now has a single responsibility (and
we might actually change it to *only* do the classpath update, not the
full synchronization in the future)

Since the update is run inside the same workspace transaction as the
import, this change does not trigger any additional builds, so there is
no performance penalty.