Let the `GradleClasspathContainerUpdater` only do a classpath update
Up until now, the `GradleClasspathContainerUpdater` did a full
synchronization on the Gradle project. The reasoning behind this was
that projects would automatically get updated on startup. This approach
had drawbacks though:

- it mixes concerns (classpath update and full project sync)
- it only works for Java projects, which could confuse users
- it adds a lot of background work when Eclipse is started

We decided to get rid of the synchronization on startup completely. This
is consistent with the behavior in IntelliJ, where the user decides when
he wants to synchronize.

Right now, we still need to run the classpath updater on startup, so the
performance overhead is still there. This will be mitigated once we
implement model persistence.