Use raw IO to read project config as a fallback
If a Buildship project is imprted with the 'Import existing project
into workspace' wizard, then Buildship throws an exception saying
that the project configuration can't be read.

The root cause is that the project creation event is fired before the
content is refreshed.

Similarly to JDT core, we fall back to raw file operations to read
the configurtion in this scenario.