Improve how we hide the Gradle build folder
Up until now we added a resource filter to the project which would
exclude the build folder from the Eclipse file system completely. The
upside is that the user does not see the build folder and its contents
don't show up in the 'Open Resource' dialog. The downside is that none
of the contents are available to other plugins like JDT. So if the user
has generated sources in the build folder, such projects would be broken
by Buildship

The following changes were made to fix this problem:

- We no longer add a resource filter. This means that e.g. folders under
the build folder can be marked as source folders and work with JDT as
expected.
- We mark the build folder as 'derived', so that it does not show up in
the 'Open Resource' dialog by default
- We add a viewer filter contribution to the Navigator, Project Explorer
and Package Explorer. This allows the user to decide for himself whether
he wants to see the build folder in those views.