Improve test performance by only waiting for Gradle jobs
Pooling the idle flag on the job manager leads to much longer test times
than necessary, as the job manager is almost always busy. For instance,
the JDT index thread reschedules itself with a delay of 0ms until it has
finally finished indexing the project's classpath. This means that all
our Java project tests waited for the JRE to be indexed again and again.

Replaced this pattern by joining an a Gradle job family.