Split builder test for normal and broken projects
This is an experiment in how we could group our tests for better
readability. Having classes that focus on a single scenario allows us to
put most of the setup logic in a common setup method, making the
individual tests shorter and easier to follow.

In this specific case there where two scenarios: A normal project and a
broken project. Some of the methods in the class were only used by the
normal project tests, some only by the broken project tests. Since they
did not share the same setup, each test method had to repeat the setup
code. By splitting the two scenarios, we get the benefits listed above.

Of course this is only a very simple example, but this it what makes it
instructive in my opinion: Even for already simple classes we can still
improve by splitting.