Always refresh the project up-front instead of implicitly
Up until now the setting of the resource filters triggered an implicit
refresh on the project. This has been moved up and made explicit. All
the other locations have been set to background_refresh, so we don't do
duplicate work.

We should get rid of the need to refresh the resource tree, as that
lowers the perceived performance of Buildship. However, we currently
need it to mark the build and .gradle folder as derived.

The correct way to achieve the above would be to have an
IResourceChagneListener that listens for the .gradle and build folder to
appear and setting them to derived then.