Reduce API surface of WorkspaceGradleOperations
A caller should never be allowed to synchronize only on project of a
build. This can leave the workspace in an inconsistent state.
Consequently, I reduced the API to the one method that takes the build
as an argument and inlined the JavaDoc.

The two removed methods were not called anywhere in the code base.