Remove unused Gradle nature context
The ContextActivatingSelectionListener was called for any selection that
the user made anywhere in the workbench. If that selection contained a
Gradle project, the "Gradle project selected" context was set. This was
only used in once place - for checking whether to run a synchronize
during a refresh (F5) operation. 

In fact, that logic works just fine without the check, because the
ProjectSynchronizer does the same check again based on the
ExecutionEvent. If it finds no projects, it is a no-op.

The ContextActivatingSelectionListener was the source of many automated
error reports, mostly because of race conditions. For instance, it would
throw exceptions when the user deleted a project while the listener was
inspecting it.

Removing it makes the code base simpler, removes an unnecessary
performance penalty and will reduce the amount of error reports.