Access markers in a workspace transaction
It seems that there is a background process running which changes the
markers while we are trying to read them. It might be JDT's classpath
validator. Running in a workspace transaction will at least guarantee us
to see a consistent state. We might see the state before the marker is
removed (in which case we would at least know what marker is causing our
problems) or after the marker is removed (which would be what we want).