Run workspace sync in a Java model batch operation
This has two benefits:

1. JDT will only validate the projects once we are done with all
modifications instead of on each change.

2. JDT will not send any events to listeners. This was especially nasty,
because it meant that projects became visible while our operation was
still running. As a result, the tasks view was updating itself several
times during a sync operation. After the change, there will only be one
atomic event at the end of the operation.