Don't fire VM notifications while it is not fully configured
I noticed that when we add a new VM, Eclipse logs an error that the
location cannot be found. This is because we were using a real VM
install and modifying it with each method call, creating events in the
process.

We now use a stand-in mechanism that allows us to configure the VM and
then create it in one atomic operation.