Keep user-defined settings on source folders
The Tooling API does not provide many of the important attributes that a
source folder can have. This causes a lot of trouble for users who rely
on them, as we currently overwrite them on overy refresh.

Until the TAPI provides the necessary information (and it is
configurable in the Gradle build), we will keep the user-defined
attributes.