Show progress during TAPI call
The TAPI of course does not provide any estimate of the work to be done.
But we can show at least some "fake" progress by consuming a percentage
of the remaining progress bar for every tick.

This commit also shows how the use of SubMonitor drastically reduces
boilerplate. No more try-catch blocks, no messages that will never be
shown and a shorter syntax for creating children.