Use SubMonitor also in ProjectPreviewJob
This uncovered that we are actually not using the monitor there when
running the TAPI jobs. Instead we are logging part of the progress to
another monitor in the user interface.

This should be replaced with a synchronous TAPI call plus a timer task
which checks the ProgressMonitor for cancellation and cancels the token.