Don't mark sub-project folders as derived
They are not derived, we just used that as a workaround for duplicate
entries in "Open Resources". This lead to problems, because eGit for
instance automatically adds derived folders to .gitignore.

It turns out that there is an option to "Filter duplicate resources", it
is just not on by default and it does not work well enough together with
the "Filter derived resources" option. These problems are tracked by
https://bugs.eclipse.org/bugs/show_bug.cgi?id=490899