Resynchronize the workspace when projects are added/removed
In order for name-deduplication and dependency substitution to stay
up-to-date, we need to re-synchronize the composite when the user
manually adds or removes projects.

This logic was already present in the task view and has been moved into
the core plugin. The task view now simply listens for a
Buildship-specific event sent by the core plugin.