Fix flakey test
Using IWorkspace#run() did not have the desired effect here. Instead of
publishing all notifications synchronously, it published them in a
background job, which made the test even more likely to fail. 

The reason for this behavior is what seems like a Bug in the WorkManager
to me: The preparedOperationCount is in a shared variable instead of a
ThreadLocal. So any fast-running jobs scheduled inside the runnable will
lead to its preparedOperations count being greater than 0 when we leave
the top-level operation.

Using this change, we no longer try holding back the notifications, but
instead hold the workspace lock until we are ready for the
synchronization job to run.