Show partial task view content if some projects are broken
Allow partial failure when displaying the tasks view

Instead of fetching an all-or-nothing `OmniEclipseWorkspace` model, we
use the `ModelResults` class from the tooling API. This way consumers
can decide whether they want to continue or fail when some models could
not be fetched. The tasks view should clearly continue and show all
intact projects. The synchronization on the other hand should fail, so
it doesn't put the workspace into an inconsistent state.