Unresolved extension point is marked as warning
The branding plugin contains an extension defined in Mylyn which is 
not part of the target platform. As a result, the workspace contains
an error marker after a successful workspace setup.

To leverage that without modifying the target platform (and effectively
making Buildship a +3 project) we change the corresponding marker from
error to warning.