Show "Refresh Gradle Projects" on .gradle files
Until now we only showed the refresh action on the Gradle projects
themselves. This is cumbersome when the user is editing a .gradle file
and his focus is on the editor. Also, the prominent green G on .gradle
files invites users to right-click on them, but until now there were no
Gradle-specific actions on them. 

This adds the refresh action to all .gradle files both in the explorer
as well as in the editor.