Add "Add Gradle Nature" to "Configure" menu.
This allows users to easily convert a plain Eclipse project to a
Buildship project. This is equivalent to opening the import wizard,
browsing to the project's physical location and hitting 'Finish'.

This is needed quite often when users migrate from existing tools like
STS or when they use SVN/Git, which currently don't support the Gradle
Import wizard directly.