Make Gradle User Home a workspace-level preference
The option was already removed from the import wizard and is also no
longer available in the run configurations. Instead it can be set in a
workspace preference.

I tried to keep the code changes minimal for now. We'll want to refactor
the whole configuration logic at some point. For instance, javaHome,
jvmArgs and gradleArgs are currently on a class called
"FixedRequestAttributes", even though these can be different from
request to request without creating a new connection. The onyl things
that require a new connection are Project directory, Gradle distribution
and Gradle user home.