Avoid missing task view updates.
We were trying to avoid unnecessary task view updates based on the
assumption that "if an update is already running, then we don't need
another one". This is not true if the workspace is rapidly changing like
in our tests.

There is only two ways to have consistency here: Either run every task
view reload job or give them a workspace lock and de-duplicate them. The
latter would block the user for a low-priority operation, which is not
acceptable. So we go with the former solution.