Add natures in a batch operation instead of one-by-one
Eclipse will validate the consistency of the project description when it
is updated. By setting them one-by-one there can easily be situations
where we add a nature that is missing a prerequisite nature that comes
later in the model.

This has been reported by several users who failed to import Groovy
projects because Buildship tried to add the Groovy nature before adding
the Java nature