Handle AggregateException correctly.
Recursively analyze the exceptions contained in the aggregate and inform
the user/log the exceptions as appropriate. Return an aggregate status
that has the maximum severity of all statuses combined.

This makes sure we don't log an error when the AggregateException just
contained cancellation exceptions or errors in the user's build script.