Handle race conditions in resource delete events gracefully
When the user deletes a build folder, there is a race condition between
the UI update calling `isBuildFolder` and the build folder disappearing
from the file system. Instead of throwing an exception, we now log this
fact and treat it as a normal condition.

The alternative would be to synchronize access to the filesystem, which
would greatly degrade performance, as this code is called every time the
content of the Project Explorer is refreshed.