Use persistent property to remember linked resources
Writing them to the .prefs file is confusing our users. They think that
they can customize them or that they are somehow important. In fact,
they are just internal state which doesn't make sense to expose to the
user and also should not be checked into version control.

Linked resources are workspace metadata, so persistent properties are a
good fit, since their lifecycle is also bound to the workspace.

The size limitation of persistent properties (2KB) should be acceptable,
since we store relative paths, not absolute ones.

If we actually encounter a real world project with so many linked
resources that this becomes an issue, we can use a custom file inside
our plugin's state location, similar to our classpath cache.