Temporarily disable custom Java home for test Eclipse
Currently both the Buildship 1.0.x and master branch are sharing
the same CI builds. This means that Java 6 is used for older target
platform to execute the integration tests. Since the master branch
is using Gradle Tooling API 3.0 the builds fail.

Once we change the execution environment to Java 7 and adjust the
CI, then the custom java home location should be re-enabled.