Fix classpath container ordering
ClasspathContainerUpdater deleted the JRE container from the classpath
if it was part of the model and then re-added it at the end. This
changed the ordering of the classpath during synchronization and
violated the assumption that the JRE container always comes after the
Gradle classpath container.

The problem has been fixed in a following way: all containers removed or
added from or to the classpath first removed from their original place,
then the added entries are placed after the source folders (with the 
ordering preserved).