Reduce scope of workspace and Java model transactions
Doing the whole synchronization in a single transaction lead to lots of
memory pressure in bigger projects. We originally introduced this
transaction because the tasks view could not handle rapid updates. This
has been fixed in the meantime and the view now behaves sanely even when
there are hundreds of "project added" events in short succession.