BuildCommandUpdater adds GradleProjectBuilder by default
After the latest project synchronization changes the Gradle builder
was no longer added to the projects. This was due to the fact that we
relied on the Gradle nature extension point to add the builder along
with the nature. The builders were overridden by the BuildCommandUpdater
when the used Gradle versions provided natures and builders.

This commit fixes the problem by applying the Gradle builder by default
on all Gradle projects. This behaviour is symmetric with the
ProjectNatureUpdater class implementation.