Make Gradle classpath containers of different projects unequal
JDT de-duplicates classpath containers when computing the runtime
classpath for a project. By default all containers with the same path
are considered equal. This breaks when a user manually adds a Gradle
project to the classpath of another Gradle project (which is still a
valid use case while composite builds and WTP don't work together, as
well as for quick experimentation). See
https://github.com/gradle/gradle/issues/1516#issuecomment-284649135

The fix is simple: Our container should be considered different for each
project.

Signed-off-by: Stefan Oehme <st.oehme@gmail.com>