GradleRunConfigurationAttributes can be created from empty launch configuration
When creating a new run configuration in Eclipse, the dialog is
initialized from an empty launch configuration. In that case the
GradleRunConfigurationAttributes class tried to use an empty
string as a default for the the Gradle distribution which
led to a serialization exception. This commit adjusts the default
Gradle distribution value to prevent the problem.