Fix loading build actions in development mode
ConnectionAwareLauncherProxy.ideFriendlyCompositeModelQuery() loads the
CompositeModelQuery build action class in a separate classloader when
Eclipse is launched in development mode because PDE returns wrong class
URLs for this usecase.

However, the custom URL class loader should not be closed after the
loading the build action as it prevents the Gradle daemon to load the
same class upon the TAPI invocation.

This commit removes the problematic class loader closing snippet.