Don't show error dialog when GradleConnectionException is received
The build launcher proxies used for executing builds produce
InvocationTargetException when a GradleConnectionException is
received from the TAPI. This exception is unknown to the
ToolingApiInvoker, therefore a general error dialog is displayed
for the user.

To make the exception handling behave correctly, this commit safely
unwraps and rethrows the root cause of the InvocationTargetException.