Further improve Executions view performance
The Executions view became really slow when a real-life build is executed
(e.g. sanityCheck on the gradle/gradle project). This is due to the fact
that all labels in the viewer was refreshed every 100 miliiseconts.

To mitigate the problem, this commit takes only the operation items that
are being executed in the build, and refreshes their label only.

A minor extra enhancement in this commit is that the execution tree now
uses `setUseHashlookup(true)` which makes sure that an object lookup
- used by TreeViewer.update(operationItem) - is performant on larger
input too.