Executions view does not expand everything on input change
The Executions view used to expand all nodes for every operation
descriptor. This made the view practically unusable when the build
was running: it was impossible to close a subtree and it put quite
some pressure to the UI thread (even with the previous changes).

This commit changes the default behaviour such that only the
following nodes are expanded upon update:
- The first and second level nodes
- The nodes representing tasks
- The nodes representing failed operations