Move UI aspects out of execution progress listener
The presentation-related code and state is now handled in the 
execution page. The listener's only job is to forward events to
a job that buffers events for asynchronous UI updates. Encapsulating
the state in the UI element means that we no longer need to worry
about concurrency, except for using a thread-safe queue for buffering
the events. Everything else happens in the UI thread.