Use correct project dir when executing tasks
This commit fixes a regression introduced in Buildship 2.1. The
new preferences API ignored the project location and always
executed all tasks on the root project. As a result, task selectors
were always executed on all projects, even if they were started from
a subproject.

The fix changes the RunConfiguration interface such that it references
a project config as opposed to a build configuration.

This commit fixes https://github.com/eclipse/buildship/issues/520