Remove Maven nature when Gradle project imported
An Eclipse project should be either Gradle or Maven project but
not both at the same time. If both nature were present, the plugins
would compete on how to calculate the classpath for the associated
run configurations. To resolve this situation, Buildship now
actively removes the Maven nature during the Gradle synchronization.

This commit fixes #617