Fix flaky ImportingSymlinkedProject
In Eclipse 4.2 the test sporadically fails because of the root
project directory is missing from the project properties.
I assumed that this is due to a synchronization issue where the
project properties are changed twice in rapid succession. I can't
reproduce the problem locally, nor can I find a report in Bugzilla.
The fix is swap the order of saving the build and the project
configuration. This change is transparent, yet it seems that the
flakiness is gone.