Viewer filter performance optimization
Viewer filter performance is essential for the Common Navigator (i.e.
package explorer) to remain responsive. Therefore test for folders using
simple instanceof instead of Adapters, which can load and initialize a
huge amount of classes as side effect. In
https://bugs.eclipse.org/bugs/show_bug.cgi?id=539470 there is an
example, where this filter was currently evaluated while eclipse was not
responsive.

I cannot guarantee that this change fixes the mentioned performance
issue (since I cannot reproduce it locally), but it is definitely a
performance improvement.

Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>