Add missing options to workspace and project configuration (#810)
This pull request lets the user configure the following properties in the workspace and build configuration:
- Java home
- program arguments
- JVM arguments
- show console view
- show executions view

To implement that, I've added the UI controls to the project preference page that is reused by the workspace, project, and run configuration dialogs. Consequently, I've removed arguments and java home tabs from the run configuration dialog as it provided duplicate functionality.

I've also adjusted the project import as it now respects the show console view configuration.