Always populate Executions view when it's opened
Before this commit the executions view was only populated when the
`Show executions view` preference was enabled. This was not correct
as the preference should only make the view visible on each Gradle
invocation. The correct behavior is to populate the view on each
incocation when visible on the UI.