Always apply the eclipse plugin via an init-script
Gradle now applies the eclipse plugin earlier. This causes the eclipse
and wtp tasks to appear in the task list even when the project does not
apply the eclipse plugin in the build script. This causes issues down the
line such as failures during wtp project synchronization. We unconditionally
apply the eclipse plugin in a build script to avoid these issues.

Signed-off-by: Christian Fraenkel <christian-ec@fraenkel.net>