Fix broken import when project already defines Gradle classpath container
GradleBuild has a static map keeping tabs on the currently running
synchronization, where the map keys are the GradleBuild instances.
Consequently, calling isSynchronizing() will return true even though a
synchronization is already running with different attributes.

This problem occurs in the following use-case:
- The project being imported already contains the Gradle classpath
  container
- During the import, the user specifies custom import settings.
If the classpath container is present, its initializer tries to check
if the Gradle build is already synchronizing before scheduling a
synchronization. The initializer, however, reads the import properties
from the disk which defaults to inheriting workspace defaults.

The fix the issue, the running synchronization operations are tracked
by the root project location.