Add failure condition to detect potential credentials leaks (#1013)
Uses a fake unmasked credentials value to fail the build. This fake value is added in TC build parameters and environment variables. It won't detect every credentials leak, but will at least fail the build if all the environment variables or build properties (that might include sensitive values) are dumped during the build.

Signed-off-by: Vaidotas Valuckas <vaidotas.valuckas@gmail.com>