Work around caching init scripts with compiled with unsupported Java version
In older Gradle versions, the compilation cache does not have the Java version for init script cache key. As a result, a build might compile and cache the init script with a higher Java version, that is incomatible with the JVM running a subsequent Gradle invocation.

To work around the problem, Buildship now adds the Java version to the script name.