Fix refreshing the tasks view on project open/close events
We observed failing tests on CI verifying that the Gradle tasks view
refreshes when workspace projects open and close. The test started to
fail for Eclipse versions 4.24 and above.

The root cause was that the event listener checked for project open
events with an equals operation:

    delta.getFlags() == IResourceDelta.OPEN

It can happen, though, that the delta contains other flags. The proper
check uses byte manipulation:

    0 != (delta.getFlags() & IResourceDelta.OPEN)

The tests started to fail only for Eclipse 4.24 because that’s the first
version that requires an encoding configured for all workspace projects.
Buildship does not do that yet (see issue #344). The missing
configuration is visualized with a warning marker that’s also observed
in the resource delta:

    delta.getFlags() == IResourceDelta.OPEN && IResourceDelta.MARKERS