Fix editor bindings for gradle.properties and Kotlin DSL files
In order for the editor content type bindings configured in the
plugin.xml to work correctly, the missing dependencies for the Generic
Text Editor have to be added.

For Gradle Properties the content type configuration should use the full
file name "gradle.properties" without a separate entry for the file
extension, so that these files are opened with the Generic Text Editor
and other property files are still handled by the Properties File
Editor.